[Hooks] Extract hook useSearchHistory
