+++ services/frontend/src/common/hooks.js
@@ -2,7 +2,6 @@ import { isEqual } from 'lodash'
 import qs from 'query-string'
 import { useCallback, useEffect, useRef, useState } from 'react'
 
-import { SEARCH_HISTORY_VERSION } from '@/constants'
 import { useGetCurriculumPeriodsQuery } from '@/redux/curriculumPeriods'
 import { useGetProgrammesQuery } from '@/redux/populations'
 import { useGetSemestersQuery } from '@/redux/semesters'
@@ -53,53 +52,6 @@ export const useTabs = (id, initialTab, { location, replace }) => {
       setTab(activeIndex)
     },
   ]
-}
-
-export const useSearchHistory = (id, capacity = 5) => {
-  const [searchHistory, setSearchHistory] = useState([])
-  const [didMount, setDidMount] = useState(false)
-
-  const getSearchHistoryStore = () => JSON.parse(localStorage.getItem('searchHistoryStore')) || {}
-
-  const saveSearchHistoryStore = newStore => localStorage.setItem('searchHistoryStore', JSON.stringify(newStore))
-
-  const saveSearchHistory = () => {
-    const searchHistoryStore = getSearchHistoryStore()
-    searchHistoryStore[id] = searchHistory
-    saveSearchHistoryStore(searchHistoryStore)
-  }
-
-  useEffect(() => {
-    if (localStorage.getItem('searchHistoryVersion') !== SEARCH_HISTORY_VERSION) {
-      saveSearchHistoryStore({})
-      localStorage.setItem('searchHistoryVersion', SEARCH_HISTORY_VERSION)
-    }
-
-    setSearchHistory(getSearchHistoryStore()[id] || [])
-    setDidMount(true)
-  }, [])
-
-  useEffect(() => {
-    if (didMount) {
-      saveSearchHistory()
-    }
-  }, [searchHistory])
-
-  const addItem = item => {
-    const filteredSearchHistory = searchHistory.filter(sh => sh.text !== item.text)
-    setSearchHistory(
-      filteredSearchHistory.concat({ ...item, timestamp: new Date(), id: new Date().getTime() }).slice(-capacity)
-    )
-  }
-
-  const updateItem = item => {
-    const updatedSearchHistory = [{ ...item, timestamp: new Date() }].concat(
-      searchHistory.filter(s => s.id !== item.id)
-    )
-    setSearchHistory(updatedSearchHistory)
-  }
-
-  return [searchHistory, addItem, updateItem]
 }
 
 const useInterval = (callback, delay) => {
